+++ backend/property_manager/__init__.py
@@ -1,5 +1,5 @@
 """PropertyManager backend package."""
 
-APPLICATION_VERSION = "6.4.0.1"
+APPLICATION_VERSION = "6.5.0"
 API_VERSION = "v1"
 DATABASE_SCHEMA_VERSION = 1